+++ OpenGLTemplate/Pickup.h
@@ -13,7 +13,7 @@ public:
 	
 	void Initialise(CSphere* &object);
 	void Render(glutil::MatrixStack playerStack, CShaderProgram* shaderProgram, CCamera* camera);
-	void Update(float dt, const glm::vec3 &player_pos, float &score);
+	void Update(float dt, const glm::vec3 &player_pos, int &score);
 	bool Active() { return m_is_active; };
 	glm::vec3 GetPosition() const { return m_position; };
